ops-knowledge-bot-branded/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==5.0.1
tqdm==4.66.5
openai==1.43.0
langdetect==1.0.9
nltk==3.9.1